Fixed myFunction firing on button click
In plain Javascript, functions that are supposed to fire on a click or a key press or anything that is done after the page loads need to have "function" before the name of the function. I've been using so much Typescript for the past 6 months for Nudget and Boredom Busters that I forgot that Javascript needs the "function" before the name.